wtf2/requirements.txt
# FastAPI & Backend
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
xgboost==2.0.2
joblib==1.3.2

# Data Processing
geopandas==0.14.1
rasterio==1.3.9
shapely==2.0.2

# GenAI & LLM
openai==1.3.7
langchain==0.0.335
python-dotenv==1.0.0

# API & Utilities
httpx==0.25.2
python-multipart==0.0.6
aiofiles==23.2.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# GIS (Optional - if needed)
fiona==1.9.5
pyproj==3.6.1